requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==3.2.0
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
python-dotenv==0.10.3
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
